feat: 밈 상호작용 라우터 meme으로 옮김
