+++ src/controller/meme.controller.ts
@@ -7,6 +7,7 @@ import { HttpCode } from '../errors/HttpCode';
 import { CustomRequest } from '../middleware/requestedInfo';
 import { IMemeCreatePayload, IMemeUpdatePayload } from '../model/meme';
 import * as MemeService from '../service/meme.service';
+import * as UserService from '../service/user.service';
 import { logger } from '../util/logger';
 import { createSuccessResponse } from '../util/response';
 
@@ -164,12 +165,83 @@ const searchMemeByKeyword = async (req: CustomRequest, res: Response, next: Next
   }
 };
 
+const createMemeReaction = async (req: CustomRequest, res: Response, next: NextFunction) => {
+  const user = req.requestedUser;
+  const meme = req.requestedMeme;
+
+  try {
+    const result = await MemeService.createMemeInteraction(user, meme, 'reaction');
+    return res.json(createSuccessResponse(HttpCode.CREATED, 'Create Meme Reaction', result));
+  } catch (err) {
+    return next(new CustomError(err.message, err.status));
+  }
+};
+
+const createMemeSave = async (req: CustomRequest, res: Response, next: NextFunction) => {
+  const user = req.requestedUser;
+  const meme = req.requestedMeme;
+
+  try {
+    const result = await MemeService.createMemeInteraction(user, meme, 'save');
+    return res.json(createSuccessResponse(HttpCode.CREATED, 'Crate Meme Save', result));
+  } catch (err) {
+    return next(new CustomError(err.message, err.status));
+  }
+};
+
+const createMemeShare = async (req: CustomRequest, res: Response, next: NextFunction) => {
+  const user = req.requestedUser;
+  const meme = req.requestedMeme;
+
+  try {
+    const result: boolean = await MemeService.createMemeInteraction(user, meme, 'share');
+    return res.json(createSuccessResponse(HttpCode.CREATED, 'Crate Meme Share', result));
+  } catch (err) {
+    return next(new CustomError(err.message, err.status));
+  }
+};
+
+const createMemeWatch = async (req: CustomRequest, res: Response, next: NextFunction) => {
+  const user = req.requestedUser;
+  const meme = req.requestedMeme;
+
+  try {
+    // 밈 조회
+    // 최근 본 밈 추가
+    const [result, _]: [boolean, any] = await Promise.all([
+      MemeService.createMemeInteraction(user, meme, 'watch'),
+      UserService.updateLastSeenMeme(user, meme),
+    ]);
+
+    return res.json(createSuccessResponse(HttpCode.CREATED, 'Crate Meme Watch', result));
+  } catch (err) {
+    return next(new CustomError(err.message, err.status));
+  }
+};
+
+const deleteMemeSave = async (req: CustomRequest, res: Response, next: NextFunction) => {
+  const user = req.requestedUser;
+  const meme = req.requestedMeme;
+
+  try {
+    const result: boolean = await MemeService.deleteMemeSave(user, meme);
+    return res.json(createSuccessResponse(HttpCode.OK, 'Delete Meme Save', result));
+  } catch (err) {
+    return next(new CustomError(err.message, err.status));
+  }
+};
+
 export {
   getMeme,
   getTodayMemeList,
   getAllMemeList,
   createMeme,
+  createMemeSave,
+  createMemeWatch,
+  createMemeShare,
+  createMemeReaction,
   deleteMeme,
+  deleteMemeSave,
   updateMeme,
   getMemeWithKeywords,
   searchMemeByKeyword,
